# src/main/resources/openapi/openapi.yml
openapi: 3.1.0
info:
  title: proyecto_final_mitocode
  description: ''
  version: 1.0.0
tags: []
paths:
  /professionals:
    get:
      summary: Obtener Profesionales
      deprecated: false
      description: ''
      operationId: listProfessional
      tags: []
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfesionalSchemaResponse'
          headers: {}
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
    post:
      summary: Crear profesionales
      deprecated: false
      description: ''
      operationId: createProfessional
      tags: []
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfesionalSchemaRequest'
            examples: {}
        required: true
      responses:
        '201':
          description: ''
          headers: {}
        '400':
          $ref: '#/components/responses/Bad Request'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
  /professionals/{idProfessional}:
    get:
      summary: Buscar Profesionales
      deprecated: false
      description: ''
      operationId: findProfessional
      tags: []
      parameters:
        - name: idProfessional
          in: path
          description: ''
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfesionalSchemaResponse'
          headers: {}
        '404':
          $ref: '#/components/responses/Not Found'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
    put:
      summary: Editar profesionales
      deprecated: false
      description: ''
      operationId: updateProfessional
      tags: []
      parameters:
        - name: idProfessional
          in: path
          description: ''
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfesionalSchemaRequest'
            examples: {}
        required: true
      responses:
        '204':
          description: ''
          headers: {}
        '400':
          $ref: '#/components/responses/Bad Request'
          description: ''
        '404':
          $ref: '#/components/responses/Not Found'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
    delete:
      summary: Borrar Profesionales
      deprecated: false
      description: ''
      operationId: deleteProfessional
      tags: []
      parameters:
        - name: idProfessional
          in: path
          description: ''
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: ''
          headers: {}
        '404':
          $ref: '#/components/responses/Not Found'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
  /clients:
    get:
      summary: Obtener Clientes
      deprecated: false
      description: ''
      operationId: listClients
      tags: []
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClienteSchemaResponse'
          headers: {}
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
    post:
      summary: Crear clientes
      deprecated: false
      description: ''
      operationId: createClients
      tags: []
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteSchemaRequest'
            examples: {}
        required: true
      responses:
        '201':
          description: ''
          headers: {}
        '400':
          $ref: '#/components/responses/Bad Request'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
  /clients/{idClients}:
    get:
      summary: Buscar Clientes
      deprecated: false
      description: ''
      operationId: findClients
      tags: []
      parameters:
        - name: idClients
          in: path
          description: ''
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteSchemaResponse'
          headers: {}
        '404':
          $ref: '#/components/responses/Not Found'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
    put:
      summary: Editar clientes
      deprecated: false
      description: ''
      operationId: updateClients
      tags: []
      parameters:
        - name: idClients
          in: path
          description: ''
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteSchemaRequest'
            examples: {}
        required: true
      responses:
        '204':
          description: ''
          headers: {}
        '400':
          $ref: '#/components/responses/Bad Request'
          description: ''
        '404':
          $ref: '#/components/responses/Not Found'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
    delete:
      summary: Borrar Clientes
      deprecated: false
      description: ''
      operationId: deleteClients
      tags: []
      parameters:
        - name: idClients
          in: path
          description: ''
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: ''
          headers: {}
        '404':
          $ref: '#/components/responses/Not Found'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
  /bookings/professional-query:
    get:
      summary: Consultas de profesionales
      deprecated: false
      description: ''
      operationId: listProfessionalQuery
      tags: []
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservasSegunProfesionalSchemaResponse'
          headers: {}
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
  /bookings:
    get:
      summary: Obtener Reservas
      deprecated: false
      description: ''
      operationId: listBookings
      tags: []
      parameters:
        - name: idClient
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: idProfessional
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: maxDate
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: minDate
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: specialty
          in: query
          description: ''
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegistroReservaSchemaResponse'
          headers: {}
        '400':
          $ref: '#/components/responses/Bad Request'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
    post:
      summary: Crear Reservas
      deprecated: false
      description: ''
      operationId: createBookings
      tags: []
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistroReservaSchemaRequest'
            examples: {}
        required: true
      responses:
        '201':
          description: ''
          headers: {}
        '400':
          $ref: '#/components/responses/Bad Request'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
  /bookings/{idBookings}:
    get:
      summary: Buscar Reservas
      deprecated: false
      description: ''
      operationId: findBookings
      tags: []
      parameters:
        - name: idBookings
          in: path
          description: ''
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistroReservaSchemaResponse'
          headers: {}
        '404':
          $ref: '#/components/responses/Not Found'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
    put:
      summary: Editar Reservas
      deprecated: false
      description: ''
      operationId: updateBookings
      tags: []
      parameters:
        - name: idBookings
          in: path
          description: ''
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistroReservaSchemaRequest'
            examples: {}
        required: true
      responses:
        '204':
          description: ''
          headers: {}
        '400':
          $ref: '#/components/responses/Bad Request'
          description: ''
        '404':
          $ref: '#/components/responses/Not Found'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
    delete:
      summary: Borrar Reservas
      deprecated: false
      description: ''
      operationId: deleteBookings
      tags: []
      parameters:
        - name: idBookings
          in: path
          description: ''
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: ''
          headers: {}
        '404':
          $ref: '#/components/responses/Not Found'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
  /availability-hours:
    get:
      summary: Obtener Horarios de Disponibilidad
      deprecated: false
      description: ''
      operationId: listAvailabilityHour
      tags: []
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HorarioDisponibleSchemaResponse'
          headers: {}
        '400':
          $ref: '#/components/responses/Bad Request'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
    post:
      summary: Crear Horarios de Disponibilidad
      deprecated: false
      description: ''
      operationId: createAvailabilityHour
      tags: []
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HorarioDisponibleSchemaRequest'
            examples: {}
        required: true
      responses:
        '201':
          description: ''
          headers: {}
        '400':
          $ref: '#/components/responses/Bad Request'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
  /availability-hours/{availabilityHoursId}:
    get:
      summary: Buscar Horarios de Disponibilidad
      deprecated: false
      description: ''
      operationId: findAvailabilityHour
      tags: []
      parameters:
        - name: availabilityHoursId
          in: path
          description: ''
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HorarioDisponibleSchemaResponse'
          headers: {}
        '404':
          $ref: '#/components/responses/Not Found'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
    put:
      summary: Editar Horarios de Disponibilidad
      deprecated: false
      description: ''
      operationId: updateAvailabilityHour
      tags: []
      parameters:
        - name: availabilityHoursId
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HorarioDisponibleSchemaRequest'
            examples: {}
        required: true
      responses:
        '204':
          description: ''
          headers: {}
        '400':
          $ref: '#/components/responses/Bad Request'
          description: ''
        '404':
          $ref: '#/components/responses/Not Found'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
    delete:
      summary: Borrar Horarios de Disponibilidad
      deprecated: false
      description: ''
      operationId: deleteAvailabilityHour
      tags: []
      parameters:
        - name: availabilityHoursId
          in: path
          description: ''
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: ''
          headers: {}
        '404':
          $ref: '#/components/responses/Not Found'
          description: ''
        '500':
          $ref: '#/components/responses/Internal Error'
          description: ''
      security: []
webhooks: {}
components:
  schemas:
    ProfesionalSchemaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombres:
          type: string
        apellidos:
          type: string
        especialidad:
          type: string
      required:
        - id
        - nombres
        - apellidos
        - especialidad
    ProfesionalSchemaRequest:
      type: object
      properties:
        nombres:
          type: string
        apellidos:
          type: string
        especialidad:
          type: string
      required:
        - apellidos
        - especialidad
        - nombres
    ClienteSchemaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombres:
          type: string
        apellidos:
          type: string
        email:
          type: string
        telefono:
          type: string
      required:
        - id
        - nombres
        - apellidos
        - email
        - telefono
    ClienteSchemaRequest:
      type: object
      properties:
        nombres:
          type: string
        apellidos:
          type: string
        email:
          type: string
        telefono:
          type: string
      required:
        - nombres
        - apellidos
        - email
        - telefono
    HorarioDisponibleSchemaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        profesional:
          $ref: '#/components/schemas/ProfesionalSchemaResponse'
        fecha:
          type: string
          format: date
        horaInicio:
          type: string
          format: date-time
        horaFin:
          type: string
          format: date-time
      required:
        - id
        - profesional
        - fecha
        - horaInicio
        - horaFin
    HorarioDisponibleSchemaRequest:
      type: object
      properties:
        profesionalId:
          type: string
        fecha:
          type: string
          format: date
        horaInicio:
          type: string
          format: date-time
        horaFin:
          type: string
          format: date-time
      required:
        - profesionalId
        - fecha
        - horaInicio
        - horaFin
    RegistroReservaSchemaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fecha:
          type: string
          format: date
        horaInicio:
          type: string
          format: date-time
        horaFin:
          type: string
          format: date-time
        cliente:
          $ref: '#/components/schemas/ClienteSchemaResponse'
        profesional:
          $ref: '#/components/schemas/ProfesionalSchemaResponse'
        estado:
          type: string
          enum:
            - CREADA
            - CANCELADA
            - COMPLETADA
      required:
        - id
        - fecha
        - horaInicio
        - horaFin
        - cliente
        - profesional
        - estado
    RegistroReservaSchemaRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fecha:
          type: string
          format: date
        horaInicio:
          type: string
          format: date-time
        horaFin:
          type: string
          format: date-time
        clienteId:
          type: string
          format: uuid
        profesionalId:
          type: string
          format: uuid
      required:
        - id
        - fecha
        - horaInicio
        - horaFin
        - clienteId
        - profesionalId
    ReservaSegunProfesionalSchemaResponse:
      type: object
      properties:
        cliente:
          type: string
        profesional:
          type: string
      required:
        - cliente
        - profesional
    ReservasSegunProfesionalSchemaResponse:
      type: object
      properties: {}
      additionalProperties:
        $ref: '#/components/schemas/ReservaSegunProfesionalSchemaResponse'
  responses:
    Not Found:
      description: ''
      content:
        application/json:
          schema:
            title: ''
            type: object
            properties:
              description:
                type: string
            required:
              - description
    Internal Error:
      description: ''
      content:
        application/json:
          schema:
            title: ''
            type: object
            properties:
              description:
                type: string
            required:
              - description
    Bad Request:
      description: ''
      content:
        application/json:
          schema:
            title: ''
            type: object
            properties:
              description:
                type: string
              errors:
                type: array
                items:
                  type: string
            required:
              - description
              - errors
    TrackingData:
      description: ''
      content:
        application/json:
          schema:
            title: ''
            type: object
            properties:
              trackingId:
                type: string
                format: uuid
              operationDate:
                type: string
                format: date-time
              01KEDHHNC9KJT1CKZVHZV5Q66Y:
                x-tmp-pending-properties: true
            required:
              - trackingId
              - operationDate
  securitySchemes: {}
servers: []
security: []